refactor cross_seed
